move cossin and atan2 into the same trig file
